requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.1
Django==3.1.7
djangorestframework==3.12.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
zope.interface==5.3.0
